A dead end.
Tried to use Boost Polygon library to compute manu contours. But that
library is only dealing with integer coordinates... Which is a serious
problem for e.g. the non-bending plane of slats where the basic pad size
 is 40/56=0.7142857143... cm
Briefly considered to represent the coordinates in some integer units
(e.g. by converting cm to micrometers), but that'd be a lot of mess just
to avoid porting the contour making code from AliRoot to O2.
Conclusion: port the contour making code to O2.